Adds 2 files with functions and imported this functions to main.js
